Fix warnings with VS on warning level 4

+++ src/appleseed.studio/mainwindow/project/disneymaterialcustomui.cpp
@@ -207,7 +207,7 @@ size_t DisneyMaterialCustomUI::find_layer_index_by_widget(const QWidget* layer_w
     }
 
     assert(!"Could not find layer widget.");
-    return ~0;
+    return ~size_t(0);
 }
 
 vector<string> DisneyMaterialCustomUI::collect_layer_names() const
