+++ src/appleseed/renderer/kernel/intersection/triangletree.cpp
@@ -966,7 +966,7 @@ void TriangleTree::store_triangles(
 
             if (leaf_size <= NodeType::MaxUserDataSize - sizeof(uint32))
             {
-                user_data_writer.write<uint32>(~0);
+                user_data_writer.write<uint32>(~uint32(0));
 
                 TriangleEncoder::encode(
                     triangle_vertex_infos,
@@ -1368,7 +1368,7 @@ bool TriangleLeafVisitor::visit(
     const uint8* user_data = &node.get_user_data<uint8>();
     const uint32 leaf_data_index = *reinterpret_cast<const uint32*>(user_data);
     const uint8* leaf_data =
-        leaf_data_index == uint32(~0)
+        leaf_data_index == ~uint32(0)
             ? user_data + sizeof(uint32)                // triangles are stored in the leaf node
             : &m_tree.m_leaf_data[leaf_data_index];     // triangles are stored in the tree
     MemoryReader reader(leaf_data);
@@ -1523,7 +1523,7 @@ bool TriangleLeafProbeVisitor::visit(
     const uint8* user_data = &node.get_user_data<uint8>();
     const uint32 leaf_data_index = *reinterpret_cast<const uint32*>(user_data);
     const uint8* leaf_data =
-        leaf_data_index == uint32(~0)
+        leaf_data_index == ~uint32(0)
             ? user_data + sizeof(uint32)                // triangles are stored in the leaf node
             : &m_tree.m_leaf_data[leaf_data_index];     // triangles are stored in the tree
     MemoryReader reader(leaf_data);
