+++ src/appleseed/renderer/kernel/lighting/sppm/sppmparameters.cpp
@@ -52,12 +52,12 @@ namespace
 {
     static size_t fixup_bounces(const int x)
     {
-        return x == -1 ? ~0 : x;
+        return x == -1 ? ~size_t(0) : x;
     }
 
     static size_t fixup_path_length(const size_t x)
     {
-        return x == 0 ? ~0 : x;
+        return x == 0 ? ~size_t(0) : x;
     }
 
     SPPMParameters::PhotonType get_photon_type(
@@ -146,8 +146,8 @@ void SPPMParameters::print() const
         "  russian roulette start bounce %s",
         pretty_uint(m_light_photon_count).c_str(),
         pretty_uint(m_env_photon_count).c_str(),
-        m_photon_tracing_max_bounces == ~0 ? "unlimited" : pretty_uint(m_photon_tracing_max_bounces).c_str(),
-        m_photon_tracing_rr_min_path_length == ~0 ? "unlimited" : pretty_uint(m_photon_tracing_rr_min_path_length).c_str());
+        m_photon_tracing_max_bounces == ~size_t(0) ? "unlimited" : pretty_uint(m_photon_tracing_max_bounces).c_str(),
+        m_photon_tracing_rr_min_path_length == ~size_t(0) ? "unlimited" : pretty_uint(m_photon_tracing_rr_min_path_length).c_str());
 
     RENDERER_LOG_INFO(
         "sppm path tracing settings:\n"
@@ -158,8 +158,8 @@ void SPPMParameters::print() const
         "  max photons per estimate      %s\n"
         "  dl light samples              %s\n"
         "  dl light threshold            %s",
-        m_path_tracing_max_bounces == ~0 ? "unlimited" : pretty_uint(m_path_tracing_max_bounces).c_str(),
-        m_path_tracing_rr_min_path_length == ~0 ? "unlimited" : pretty_uint(m_path_tracing_rr_min_path_length).c_str(),
+        m_path_tracing_max_bounces == ~size_t(0) ? "unlimited" : pretty_uint(m_path_tracing_max_bounces).c_str(),
+        m_path_tracing_rr_min_path_length == ~size_t(0) ? "unlimited" : pretty_uint(m_path_tracing_rr_min_path_length).c_str(),
         pretty_scalar(m_initial_radius_percents, 3).c_str(),
         pretty_scalar(m_alpha, 1).c_str(),
         pretty_uint(m_max_photons_per_estimate).c_str(),
